move legacy configuration types to client-js package
