+++ packages/js/client-config-builder/README.md
@@ -149,45 +149,6 @@ A complete example using all or most of the available methods.
 
 # Reference
 
-## Types
-
-```ts
-/**
- * Client configuration that can be passed to the PolywrapClient
- *
- * @remarks
- * The PolywrapClient converts the ClientConfig to a CoreClientConfig.
- */
-export interface ClientConfig {
-  /** set environmental variables for a wrapper */
-  readonly envs: Env[];
-
-  /** register interface implementations */
-  readonly interfaces: InterfaceImplementations[];
-
-  /** redirect invocations from one uri to another */
-  readonly redirects: IUriRedirect[];
-
-  /** add embedded wrappers */
-  readonly wrappers: IUriWrapper[];
-
-  /** add and configure embedded packages */
-  readonly packages: IUriPackage[];
-
-  /** customize URI resolution
-   *
-   * @remarks
-   * A UriResolverLike can be any one of:
-   *     IUriResolver<unknown>
-   *   | IUriRedirect
-   *   | IUriPackage
-   *   | IUriWrapper
-   *   | UriResolverLike[]
-   *   */
-  readonly resolvers: UriResolverLike[];
-}
-```
-
 ## ClientConfigBuilder
 
 ### Constructor
@@ -302,7 +263,9 @@ export interface ClientConfig {
    * @param uriEnvs: and object where key is the uri and value is the another object with the env variables for the uri
    * @returns IClientConfigBuilder (mutated self)
    */
-  addEnvs(uriEnvs: Record<string, Record<string, unknown>>): IClientConfigBuilder;
+  addEnvs(
+    uriEnvs: Record<string, Record<string, unknown>>
+  ): IClientConfigBuilder;
 ```
 
 ### removeEnv
@@ -565,9 +528,7 @@ export const getDefaultConfig = (): BuilderConfig => ({
     [defaultWrappers.concurrentInterface]: new Set([
       defaultPackages.concurrent,
     ]),
-    [defaultInterfaces.logger]: new Set([
-      defaultPackages.logger,
-    ]),
+    [defaultInterfaces.logger]: new Set([defaultPackages.logger]),
   },
   resolvers: [],
 });
