--- packages/js/client-config-builder/src/helpers/index.ts
@@ -1,2 +1,0 @@
-export * from "./buildPolywrapCoreClientConfig";
-export * from "./sanitizeUri";
