+++ packages/js/client-config-builder/src/index.ts
@@ -1,5 +1,4 @@
 export * from "./types";
-export * from "./helpers";
 export * from "./BaseClientConfigBuilder";
 export * from "./ClientConfigBuilder";
 export * from "./bundles";
