+++ packages/js/client-config-builder/src/types/configs/index.ts
@@ -1,4 +1,1 @@
 export * from "./BuilderConfig";
-export * from "./ClientConfig";
-export * from "./PolywrapClientConfig";
-export * from "./PolywrapCoreClientConfig";
