+++ packages/js/client/src/legacy/buildPolywrapCoreClientConfig.ts
@@ -1,12 +1,10 @@
 import {
   PolywrapClientConfig,
   PolywrapCoreClientConfig,
-  BuilderConfig,
   ClientConfig,
-} from "../types";
-import { GenericUriResolverLike } from "../types/configs/types";
-import { ClientConfigBuilder } from "../ClientConfigBuilder";
-import { sanitizeUri } from "./sanitizeUri";
+  sanitizeUri
+} from ".";
+import { GenericUriResolverLike } from "./types";
 
 import {
   CoreClientConfig,
@@ -15,6 +13,7 @@ import {
   Uri,
 } from "@polywrap/core-js";
 import { UriResolverLike as SanitizedUriResolverLike } from "@polywrap/uri-resolvers-js";
+import { BuilderConfig, ClientConfigBuilder } from "@polywrap/client-config-builder-js";
 
 export function sanitizeConfig<TUri extends Uri | string = string>(
   config: Partial<PolywrapClientConfig<TUri>> | Partial<ClientConfig>
