+++ packages/js/client/src/legacy/index.ts
@@ -1,0 +1,6 @@
+export * from './types';
+export * from './ClientConfig';
+export * from './PolywrapClientConfig';
+export * from './PolywrapCoreClientConfig';
+export * from './buildPolywrapCoreClientConfig';
+export * from './sanitizeUri';
