+++ packages/js/client/src/legacy/sanitizeUri.ts
