+++ packages/js/client/src/legacy/types/IUriPackage.ts
