+++ packages/js/client/src/legacy/types/IUriRedirect.ts
