+++ packages/js/client/src/legacy/types/IUriWrapper.ts
