+++ packages/js/client/src/legacy/types/UriResolverLike.ts
